Fix graham_number numpy.float64 serialization error
np.sqrt() returns numpy.float64 which psycopg2 cannot serialize —
it renders as np.float64(x) causing "schema np does not exist" error.
Previously harmless because bvps was always None so graham was never
computed. Now that bvps is derived, the bug surfaces.
Fix: wrap np.sqrt() result with float() before rounding.

+++ compute/engine/yearly_compute.py
@@ -562,7 +562,7 @@ def build_yearly_rows(asx_code: str, fin: pd.DataFrame,
         payout    = _clamp(_div(dps, eps) if (eps and eps > 0) else None)
         graham    = None
         if eps and bvps and eps > 0 and bvps > 0:
-            graham = round(np.sqrt(22.5 * eps * bvps), 4)
+            graham = round(float(np.sqrt(22.5 * eps * bvps)), 4)
 
         fr_yld = None
         if div_yld is not None and fpc is not None:
